wip: Implement parallel test but fails on multithreading mode
