+++ harness/Cargo.toml
@@ -25,3 +25,6 @@ tokio = { version = "1.40", features = ["full"] }
 toml = "0.8.19"
 futures = "0.3.30"
 const_format = "0.2.33"
+once_cell = "1.20.1"
+testcontainers = { version = "0.23.1" }
+integration-test-runner = { path = "./integration-test-runner" }
