+++ harness/integration-test-runner/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "integration-test-runner"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description.workspace = true
+keywords.workspace = true
+
+[dependencies]
+proc-macro2 = "1.0.87"
+quote = "1.0.37"
+syn = { version = "2.0.79", features = ["full"]}
+testcontainers = { version = "0.23.1" }
+tokio = { version = "1.40", features = ["full"] }
+
+[lib]
+proc-macro = true
