+++ harness/src/lib.rs
@@ -3,4 +3,5 @@ pub mod constant;
 pub mod logger;
 pub mod raft;
 pub mod state_machine;
+pub mod test_utils;
 pub mod utils;
