+++ raftify-cli/Cargo.toml
@@ -13,7 +13,6 @@ slog = "2"
 built = "0.5"
 clap = { version = "4.5.18", features = ["derive"] }
 raftify = { version = "=0.1.81", features = ["heed_storage", "inmemory_storage", "rocksdb_storage"] }
-rocksdb = "0.19.0"
 comfy-table = "7.1.1"
 
 [lib]
